eric-pgbm-results/requirements.txt
requests==2.31.0
flask==2.3.2
waitress==2.1.2
pyyaml==6.0.1